+++ packages/donuts/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.1.6
+
+- fix create modal error
+
+## 0.1.5
+
+- add key factory dependence
+- fix create modal error
+
 ## 0.1.4
 
 - fix bug